fix do_irq()
do_irq() will call `clear_csr(mip, MIP_MEIE)` to
clear external interrupt pending, but it is not needed.
When riscv enters trap, hart will automatically turn
off the interrupt, and when mret, it will automatically
turn on the interrupt, so we do not need to control MEIE,
and there will be no so-called interrupt nesting problem.

Signed-off-by: Chen Wang <unicorn_wang@outlook.com>